Adding Operator Packaging Instructions (#62)
## Pull Request: OLM Bundle Creation & Bug Fixes

**Summary:**
This PR gives instruction for OLM bundle creation, enabling catalog items, and includes a few bug fixes found during installation.

**How to Test:**

1. Verify catalog item creation (using the new OLM bundle) and its appearance in your OLM catalog.

2. Perform a fresh operator installation to confirm resolved issues.

   * **Tested on OpenShift:** Successfully validated on an OpenShift cluster. Screenshots demonstrating presence in Operator Hub and successful installation are available.

![image](https://github.com/user-attachments/assets/5e7cba24-3522-4457-87a3-2c36acab5e6c)

![image](https://github.com/user-attachments/assets/7df9c687-0044-4dfe-8102-5729c570bbff)

![image](https://github.com/user-attachments/assets/30fef57e-aafc-4354-a593-6e60bff4dbd9)

**Future Considerations (For Discussion):**

I was also thinking of creating some automation around creating the OLM bundle on tags, so we could start releasing it through the catalog and make it a little easier for folks to start trying out and giving feedback.


Approved-by: VaishnaviHire